pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "embodiedbench"
version = "0.1.0"
description = "Offline edge-AI benchmarking scaffold – Phase 1"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
embodiedbench = "embodiedbench.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["embodiedbench*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
